+++ src/main/java/ch/powerunit/extensions/matchers/provideprocessor/extension/ArrayContainingDSLExtension.java
@@ -45,32 +45,17 @@ public class ArrayContainingDSLExtension implements DSLExtension {
 		return new ArrrayContainsSupplier(targetName, returnType, methodName, targetMethodName).asSuppliers();
 	}
 
-	public class ArrrayContainsSupplier {
-
-		private final String targetName;
-		private String returnType;
-		private String methodName;
-		private String targetMethodName;
+	public class ArrrayContainsSupplier extends AbstractDSLExtensionSupplier {
 
 		public ArrrayContainsSupplier(String targetName, String returnType, String methodName,
 				String targetMethodName) {
-			this.targetName = targetName;
-			this.returnType = returnType;
-			this.methodName = methodName;
-			this.targetMethodName = targetMethodName;
+			super(targetName, returnType, methodName, targetMethodName);
 		}
 
+		@Override
 		public Collection<Supplier<DSLMethod>> asSuppliers() {
 			return Arrays.asList(this::generateContains1, this::generateContains2, this::generateContains3,
 					this::generateContainsN);
-		}
-
-		public String[] getOneParameter(String name) {
-			return new String[] { targetName, name };
-		}
-
-		public String getOneWith(String name) {
-			return targetMethodName + "(" + name + ")";
 		}
 
 		public DSLMethod generateContains1() {
